% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope}
\alias{rope}
\title{Rope Function}
\usage{
rope(S, lambda, Target = NULL)
}
\arguments{
\item{S}{Sample covariance matrix: symmetric p x p matrix}

\item{lambda}{Penalization parameter: non-negative scalar}

\item{Target}{(Optional) Target matrix: A symmetric and positive definite p x p matrix. The estimate approaches the target matrix as the penalty parameter increases.}
}
\value{
\item{Theta}{Estimated precision matrix}
}
\description{
Calculates explicitly a sparse precision matrix Theta using a ridge type penalty.
}
\examples{
\donttest{
set.seed(1)
n <- 10
p <- 5
Y <- array(rnorm(n*p),dim=c(n,p)) # data-matrix
S <- cov(Y) # sample covariance matrix
Theta <- rope(S=S,lambda=0.5)
}
}
\references{
M. O. Kuismin, J. T. Kemppainen & M. J. Sillanpaa (2017).
Precision Matrix Estimation With ROPE.
Journal of Computational and Graphical Statistics. https://doi.org/10.1080/10618600.2016.1278002
}
