% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Target Generating Function}
\usage{
target(
  Y = NULL,
  S = NULL,
  type,
  cor = FALSE,
  fraction = 1e-04,
  const = 1,
  safety.scaling = 1,
  nfolds.small = TRUE,
  nfolds.number = NA
)
}
\arguments{
\item{Y}{Data matrix: n x p matrix, where p is the number of predictors and n the number of observations}

\item{S}{Sample covariance matrix: symmetric p x p matrix (zero centered data)}

\item{type}{Type of the output Target. Possible values: "Identity","vI","Regression","Eigenvalue","MSC". The details to each type are in the description.}

\item{cor}{Should correlation matrices be used instead of covariance matrices. Default is set to FALSE.}

\item{fraction}{Needed for Eigenvalue type. The cutoff for which Eigenvalues are treated to be 0. Default is set to 1e-4}

\item{const}{Needed for Eigenvalue type. Default is set to 1}

\item{safety.scaling}{Needed for Regression type. Default is set to 1}

\item{nfolds.small}{Needed for Regression type. If true 10 fold is used. Default is set to TRUE.}

\item{nfolds.number}{Needed for Regression type. Number of folds used.}
}
\value{
Returns the target matrix corresponding to the input type.
}
\description{
Generates different types of targets. The following types can be used:
"Identity" The identity matrix
"vI" A scaled identity matrix, where v is 1/mean(diag(S))
"Regression" A regression based target matrix. Use parameter: safety.scaling, nfolds.small, nfolds.number to modify. Need Y for this type of target.
"Eigenvalue" An Eigenvalue based approach. Warning: In some cases the values might get to large. Use parameter: fraction, const to modify.
"MSC" Maximal Single Correlation approach.
}
\examples{
\donttest{
set.seed(1)
n <- 10
p <- 5
Y <- array(rnorm(n*p),dim=c(n,p)) # data-matrix
target <- target(Y=Y,type="Identity")
}
}
\references{
Carel F.W. Peeters,  Anders Ellern Bilgrau and Wessel N. van Wieringen (2019).
rags2ridges: Ridge Estimation of Precision Matrices from High-Dimensional Data.
CRAN. https://CRAN.R-project.org/package=rags2ridges
}
